In progress

+++ src/main/java/geo/geopointsTgBot/Bot.java
@@ -10,7 +10,9 @@ import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.json.JSONArray;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
 import java.io.File;
 import java.io.IOException;
@@ -18,11 +20,18 @@ import java.io.IOException;
 
 public class Bot extends TelegramLongPollingBot {
 
-   private Json json;
+    private Json json;
 
     public Json getJson() {
         return json;
     }
+
+    private String greeting = "Привет! Для получения kml файла с пунктами отправьте " +
+            "сообщение с географическими координатами и радиусом поиска в формате:\n" +
+            "широта, долгота, радиус поиска (например '55.168949, 61.212220, 10'";
+    private String wrongInput = "Неправильно введен запрос, убедитесь, что он соотвествует следующему формату:\n" +
+            "\"широта, долгота, радиус поиска (например '55.168949, 61.212220, 10'";
+
 
     public Bot() {
 
@@ -42,22 +51,39 @@ public class Bot extends TelegramLongPollingBot {
         return "???";
     }
 
+    public void sendText(Long who, String what) {
+        SendMessage sm = SendMessage.builder()
+                .chatId(who.toString()) //Who are we sending a message to
+                .text(what).build();    //Message content
+        try {
+            execute(sm);                        //Actually sending the message
+        } catch (TelegramApiException e) {
+            throw new RuntimeException(e);      //Any error will be printed here
+        }
+    }
+
     @Override
     public void onUpdateReceived(Update update) {
         var msg = update.getMessage();
         String chatId = String.valueOf(msg.getChatId());
+        sendText(msg.getChatId(), greeting);
 
         String[] temp = msg.getText().split(",");
-        String jsonString = "{\n\"x\":\"" + temp[0].trim() + "\",\n\"y\":\"" + temp[1].trim() + "\"\n}";
-        JSONArray jsonList = new JSONArray(json.sendJsonToUrl(jsonString));
-        File file;
-        try {
-            file = JsonToKmlTelegramSender.sendJsonListAsKmlFileInTelegram(jsonList);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
+        if (temp.length != 3)
+            sendText(msg.getChatId(), wrongInput);
+        else {
+            String jsonString = "{\n\"x\":\"" + temp[0].trim() + "\",\n\"y\":\"" + temp[1].trim() +
+                    "\",\n\"radius\":\"" + temp[2].trim() + "\"\n}";
+            JSONArray jsonList = new JSONArray(json.sendJsonToUrl(jsonString));
+            File file;
+            try {
+                file = JsonToKmlTelegramSender.sendJsonListAsKmlFileInTelegram(jsonList);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+            sendKml(file, chatId);
+            file.delete();
         }
-        sendKml(file, chatId);
-        file.delete();
     }
 
     public void sendKml(File file, String chatId) {
